requirements.txt
django-cors-headers 4.3.1
django-easyaudit-utils 0.0.6
django-easy-audit 1.3.5
django-versatileimagefield 3.0
djangorestframework 3.14.0
psycopg 3.1.13
django-rest-swagger 2.2.0
drf-spectacular 0.26.5
twilio 8.10.2
pandas 2.1.3